aider: Added a new route handler for the `/path` endpoint that sends the `mapeoConfigFolder` string as a response.
# Aider chat conversation:

USER: create a /path route that simply returns the mapeoConfigFolder string

ASSISTANT: To create a new route `/path` that returns the `mapeoConfigFolder` string, we need to modify the `api/index.js` file.

Here are the changes:

```javascript
api/index.js
<<<<<<< HEAD
  server.listen(port, () => {
    console.log(`Server running at http://${hostname}:${port}`);
  });
}

module.exports = runApp;
=======
  app.get("/path", (req, res) => {
    res.send(mapeoConfigFolder);
  });

  server.listen(port, () => {
    console.log(`Server running at http://${hostname}:${port}`);
  });
}

module.exports = runApp;
>>>>>>> updated
```

This edit block adds a new route handler for the `/path` endpoint that sends the `mapeoConfigFolder` string as a response.